+++ extra/bash-completion-next-action.sh
@@ -1,42 +1,35 @@
+#!/bin/bash
+
 _next_action()
 {
-  arguments="-a --all -b --blocked -c --config-file -d --due -f --file -h --help -n --number -o --overdue -p --priority -r --reference -s --style -t --time-travel -V --version"
+  local argument_type="all"
+  local cur=${COMP_WORDS[COMP_CWORD]}
+  local prev=${COMP_WORDS[COMP_CWORD-1]}
 
-  local cur prev
-  COMPREPLY=()
-  cur="${COMP_WORDS[COMP_CWORD]}"
-  prev="${COMP_WORDS[COMP_CWORD-1]}"
-  
-  case "${prev}" in
+  case ${prev} in
     --file|-f|--config-file|-c)
       COMPREPLY=( $(compgen -A "file" -- ${cur}) )
       return 0
       ;;
-    --priority|-p)
-      local priorities="A B C D E F G H I J K L M N O P Q R S T U V W X Y Z"
-      COMPREPLY=( $(compgen -W "${priorities}" -- ${cur}) )
-      return 0
-      ;;
-    --reference|-r)
-      local reference="always never multiple"
-      COMPREPLY=( $(compgen -W "${reference}" -- ${cur}) )
-      return 0
-      ;;
-    --style|-s)
-      local styles="abap algol algol_nu arduino autumn borland bw colorful default emacs friendly fruity igor lovelace manni monokai murphy native paraiso-dark paraiso-light pastie perldoc rainbow_dash rrt tango trac vim vs xcode"
-      COMPREPLY=( $(compgen -W "${styles}" -- ${cur}) )
-      return 0
-      ;;
-    --time-travel|-t)
-      local dates="tomorrow yesterday Monday Tuesday Wednesday Thursday Friday Saturday Sunday"
-      COMPREPLY=( $(compgen -W "${dates}" -- ${cur}) )
-      return 0
+    --priority|-p|--reference|-r|--style|-s|--time-travel|-t)
+      argument_type=${prev}
       ;;
     *)
+      case ${cur} in
+        @*|+*)
+          argument_type=${cur:0:1}
+          ;;
+        -@*|-+*)
+          argument_type=${cur:0:2}
+          ;;
+        *)
+          ;;
+      esac
       ;;
   esac
-  COMPREPLY=( $(compgen -W "${arguments}" -- ${cur}) )
+  local arguments=$(${COMP_LINE% *} --list-arguments ${argument_type//-/_} 2> /dev/null)
+  COMPREPLY=( $(compgen -W "${arguments}" -- "${cur}") )
   return 0
 }
-complete -F _next_action next-action
-
+COMP_WORDBREAKS=${COMP_WORDBREAKS/@/}  # Needed to prevent @ from being escaped
+complete -o filenames -F _next_action next-action