+++ next_action/__init__.py
@@ -2,7 +2,7 @@
 
 from .arguments import parse_arguments
 from .pick_action import next_actions
-from .output import render_next_action
+from .output import render_next_action, render_arguments
 from .todotxt import read_todotxt_files
 
 
@@ -24,5 +24,8 @@ def next_action() -> None:
         tasks = read_todotxt_files(namespace.file)
     except OSError as reason:
         parser.error(f"can't open file: {reason}")
-    actions = next_actions(tasks, namespace)
-    print(render_next_action(actions, tasks, namespace))
+    if namespace.list_arguments:
+        print(render_arguments(namespace.list_arguments, tasks))
+    else:
+        actions = next_actions(tasks, namespace)
+        print(render_next_action(actions, tasks, namespace))