+++ tests/unittests/output/test_render.py
@@ -1,14 +1,17 @@
-"""Unit tests for the render function."""
+"""Unit tests for the render functions."""
 
 import argparse
 import unittest
 
+from hypothesis import given, strategies
+from pygments.styles import get_all_styles
+
 from next_action import todotxt
-from next_action.output import render_next_action
+from next_action.output import render_next_action, render_arguments
 
 
-class RenderTest(unittest.TestCase):
-    """Unit tests for the render method."""
+class RenderNextActionTest(unittest.TestCase):
+    """Unit tests for the render next action method."""
 
     def setUp(self):
         """Set up the namespace with default arguments for all unit tests."""
@@ -73,3 +76,40 @@ class RenderTest(unittest.TestCase):
         self.assertEqual(
             "Lather before:rinse\nblocks:\n- Rinse id:rinse before:repeat\n  blocks:\n  - Repeat id:repeat",
             render_next_action([lather], [], self.namespace))
+
+
+class RenderArgumentsTest(unittest.TestCase):
+    """Unit tests for the render arguments method."""
+
+    def test_arguments(self):
+        """Test that the base arguments are rendered correctly."""
+        self.assertEqual("+ -+ --all --blocked --config-file --due --file --help --number --overdue --priority "
+                         "--reference --style --time-travel --version -@ -V -a -b -c -d -f -h -n -o -p -r -s -t @",
+                         render_arguments("all", todotxt.Tasks()))
+
+    @given(strategies.sampled_from(["__reference", "_r"]))
+    def test_reference(self, argument):
+        """Test that the reference arguments are rendered correctly."""
+        self.assertEqual("always multiple never", render_arguments(argument, todotxt.Tasks()))
+
+    @given(strategies.sampled_from(["__time_travel", "_t"]))
+    def test_time_travel(self, argument):
+        """Test that the time-travel arguments are rendered correctly."""
+        self.assertEqual("tomorrow yesterday Monday Tuesday Wednesday Thursday Friday Saturday Sunday",
+                         render_arguments(argument, todotxt.Tasks()))
+
+    @given(strategies.sampled_from(["__style", "_s"]))
+    def test_style(self, argument):
+        """Test that the style arguments are rendered correctly."""
+        self.assertEqual(" ".join(sorted(get_all_styles())), render_arguments(argument, todotxt.Tasks()))
+
+    @given(strategies.sampled_from(["__priority", "_p"]))
+    def test_priority_no_prios(self, argument):
+        """Test that the priority arguments are rendered correctly."""
+        self.assertEqual("", render_arguments(argument, todotxt.Tasks()))
+
+    @given(strategies.sampled_from(["__priority", "_p"]))
+    def test_priority(self, argument):
+        """Test that the priority arguments are rendered correctly."""
+        self.assertEqual("A C",
+                         render_arguments(argument, todotxt.Tasks([todotxt.Task("(A) A"), todotxt.Task("(C) C")])))