+++ tests/unittests/test_cli.py
@@ -201,3 +201,48 @@ context or project is mistaken for an argument to an option.
         self.assertEqual(
             [call("Nothing to do! (warning: unknown projects: AutumnCleaning, SpringCleaning)"), call("\n")],
             mock_stdout_write.call_args_list)
+
+    @patch.object(sys, "argv", ["next-action", "--list-arguments", "@"])
+    @patch("fileinput.open", mock_open(read_data="\nWalk the dog @park\nWrite proposal +NewProject\n"))
+    @patch.object(sys.stdout, "write")
+    def test_list_contexts(self, mock_stdout_write):
+        """Test that the contexts are listed."""
+        next_action()
+        self.assertEqual(
+            [call("@park"), call("\n")], mock_stdout_write.call_args_list)
+
+    @patch.object(sys, "argv", ["next-action", "--list-arguments", "+"])
+    @patch("fileinput.open", mock_open(read_data="\nWalk the dog @park\nWrite proposal +NewProject\n"))
+    @patch.object(sys.stdout, "write")
+    def test_list_projects(self, mock_stdout_write):
+        """Test that the projects are listed."""
+        next_action()
+        self.assertEqual(
+            [call("+NewProject"), call("\n")], mock_stdout_write.call_args_list)
+
+    @patch.object(sys, "argv", ["next-action", "--list-arguments", "_@"])
+    @patch("fileinput.open", mock_open(read_data="\nWalk the dog @park\nWrite proposal +NewProject\n"))
+    @patch.object(sys.stdout, "write")
+    def test_list_excluded_contexts(self, mock_stdout_write):
+        """Test that the excluded contexts are listed."""
+        next_action()
+        self.assertEqual(
+            [call("-@park"), call("\n")], mock_stdout_write.call_args_list)
+
+    @patch.object(sys, "argv", ["next-action", "--list-arguments", "_+"])
+    @patch("fileinput.open", mock_open(read_data="\nWalk the dog @park\nWrite proposal +NewProject\n"))
+    @patch.object(sys.stdout, "write")
+    def test_list_excluded_projects(self, mock_stdout_write):
+        """Test that the excluded projects are listed."""
+        next_action()
+        self.assertEqual(
+            [call("-+NewProject"), call("\n")], mock_stdout_write.call_args_list)
+
+    @patch.object(sys, "argv", ["next-action", "--list-arguments", "__priority"])
+    @patch("fileinput.open", mock_open(read_data="\nWalk the dog @park\n(A) Write proposal\n(C) Get permit"))
+    @patch.object(sys.stdout, "write")
+    def test_list_priorities(self, mock_stdout_write):
+        """Test that the priorities are listed."""
+        next_action()
+        self.assertEqual(
+            [call("A C"), call("\n")], mock_stdout_write.call_args_list)