+++ tests/unittests/todotxt/test_tasks.py
@@ -52,3 +52,23 @@ class ProjectsTest(unittest.TestCase):
     def test_tasks_with_projects(self):
         """Test that a task collection with tasks with projects has those projects."""
         self.assertEqual({"Foo", "Bar", "Baz"}, Tasks([Task("Todo +Foo +Bar"), Task("Todo +Foo +Baz")]).projects())
+
+
+class PrioritiesTest(unittest.TestCase):
+    """Unit tests for the priorities method."""
+
+    def test_no_tasks(self):
+        """Test that an empty tasks collection has no priorities."""
+        self.assertEqual(set(), Tasks().priorities())
+
+    def test_task_without_priority(self):
+        """Test that a task collection with tasks without priorities has not priorities."""
+        self.assertEqual(set(), Tasks([Task("Todo")]).priorities())
+
+    def test_task_with_priority(self):
+        """Test that a task collection with a task with a priority has that priority."""
+        self.assertEqual(set(["C"]), Tasks([Task("(C) Todo")]).priorities())
+
+    def test_tasks_with_priorities(self):
+        """Test that a task collection with tasks with priorities has those priorities."""
+        self.assertEqual(set(["X", "Y"]), Tasks([Task("(X) Todo"), Task("(Y) Todo")]).priorities())